Remove unneeded import and require for CSS and JS files
Every JS asset in the 'javascripts' folder and CSS assets imported from applications.scss
with '@import' are automatically loaded.
However, they are mixed together and loaded as a whole for every page. It is possible to
modify this behavior, but I don't know if it is a good idea.
Thus I let style.min.css loaded manually with a link tag for now, since it is only used
for one view.