fix breaking bug for some courses index views
It fixes a bug introduced in dccd41f2eeff94d58a4bfac89bd6a2a94aa0a154.
That commit was just there to order courses more nicely in
courses index views but it wasn't simple enough of a fix to be pushed
that late and so it slipped through our last minute testing.
Courses index also work fine after reverting that commit (they are just
not in a nice order), but this commit is a better solution.